+++ app/src/main/java/jp/ac/it_college/stds/androidsteganography/scene/DecodeScene.kt
@@ -1,11 +1,6 @@
 package jp.ac.it_college.stds.androidsteganography.scene
 
-import android.content.ContentValues
-import android.content.Context
-import android.content.SharedPreferences
 import android.graphics.Bitmap
-import android.os.Environment
-import android.provider.MediaStore
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -32,15 +27,9 @@ import jp.ac.it_college.stds.androidsteganography.ui.theme.AndroidSteganographyT
 import androidx.compose.ui.platform.LocalContext
 import jp.ac.it_college.stds.androidsteganography.components.common.SimpleLSB
 import jp.ac.it_college.stds.androidsteganography.components.common.readSteganography
+import jp.ac.it_college.stds.androidsteganography.components.fileOperations.saveAndExtractZip
+
 import kotlinx.coroutines.launch
-import java.io.BufferedOutputStream
-import java.io.File
-import java.io.FileOutputStream
-import java.util.zip.ZipEntry
-import java.util.zip.ZipInputStream
-
-
-lateinit var prefSetting: SharedPreferences
 
 @Composable
 fun DecodeScene(modifier: Modifier = Modifier, onDecodeClick: () -> Unit = {}, decReceive: Bitmap) {
@@ -48,14 +37,11 @@ fun DecodeScene(modifier: Modifier = Modifier, onDecodeClick: () -> Unit = {}, d
     val context = LocalContext.current
     val scope = rememberCoroutineScope()
     var zipBitList: MutableList<Int> = remember { mutableListOf<Int>().toMutableList() }
-    var kara: ByteArray? by remember { mutableStateOf(null) }
-    var fileName: String by remember {
-        mutableStateOf("")
-    }
+    var byteArray: ByteArray? by remember { mutableStateOf(null) }
     var text by remember { mutableStateOf("") }
-
-
-    println("kara: $kara")
+    var bool: Boolean by remember {
+        mutableStateOf(false)
+    }
 
     Column(
         modifier = Modifier.fillMaxSize()
@@ -84,63 +70,24 @@ fun DecodeScene(modifier: Modifier = Modifier, onDecodeClick: () -> Unit = {}, d
             )
         }
 
-        Button(onClick = { kara = decReceive.readSteganography(zipBitList, (SimpleLSB::read)) }) {
+        Button(onClick = {
+            byteArray = decReceive.readSteganography(zipBitList, (SimpleLSB::read))
+            bool = !bool
+        }) {
             Text(text = "読み込み")
         }
-//        fileName = "example24.zip"
         Button(onClick = {
             scope.launch {
-                saveAndExtractZip(context, kara!!, "${text}.zip", "extracted_files")
+                saveAndExtractZip(context, byteArray!!, "${text}.zip", "extracted_files")
             }
-        }) {
+        },
+            enabled = bool
+        ) {
             Text(text = "保存")
         }
         Button(onClick = onDecodeClick) {
             Text(text = "終了")
         }
-    }
-}
-
-
-fun saveAndExtractZip(context: Context, byteArray: ByteArray, zipFileName: String, extractToDirName: String) {
-    try {
-        val values = ContentValues().apply {
-            put(MediaStore.MediaColumns.DISPLAY_NAME, zipFileName)
-            put(MediaStore.MediaColumns.MIME_TYPE, "application/zip")
-            put(MediaStore.MediaColumns.RELATIVE_PATH, Environment.DIRECTORY_DOWNLOADS)
-        }
-
-        val resolver = context.contentResolver
-        val uri = resolver.insert(MediaStore.Downloads.EXTERNAL_CONTENT_URI, values)
-            ?: throw Exception("ファイル作成に失敗")
-
-        resolver.openOutputStream(uri).use { outputStream ->
-            outputStream?.write(byteArray)
-        }
-
-        val zipInputStream = ZipInputStream(byteArray.inputStream())
-        val downloadsDir = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)
-        val extractDir = File(downloadsDir, extractToDirName)
-
-        var zipEntry: ZipEntry? = zipInputStream.nextEntry
-        while (zipEntry != null) {
-            val file = File(extractDir, zipEntry.name)
-            if (zipEntry.isDirectory) {
-                file.mkdirs()
-            } else {
-                file.parentFile?.mkdirs()
-                FileOutputStream(file).use { fileOutputStream ->
-                    BufferedOutputStream(fileOutputStream).use { bufferedOutputStream ->
-                        zipInputStream.copyTo(bufferedOutputStream)
-                    }
-                }
-            }
-            zipEntry = zipInputStream.nextEntry
-        }
-        zipInputStream.closeEntry()
-        zipInputStream.close()
-    } catch (e: Exception) {
-        e.printStackTrace()
     }
 }
